Replaced `TA_Edicion_Impresa::title` with the edicion impresa date.
Added articles list to the content of an edicion impresa.

Removed the `the_content` filter from single articles, and applied directly
in the `TA_Article_Data::content` property.

Filtered the title of the site when on a single edicion impresa page.

Use of header types in container with header block.

Removed unused markup fron `articulo-simple.php`.