Articles block uses balancer offset to prevent showing articles that had been already shown.
The content of `TA_Article` and `TA_Edicion_Impresa` is not longer being
filtered through the `the_content` filter, to avoid executing unnecessary scripts
that could have side effects, even the `::get_the_content()` was not being directly called.

In `get_ta_articles_from_query` the pupulation of the article instance was made
optional, and can be indicated through the second argument `$args`.

In the articles block, when balanced, a verification was added to `$balancer_articles_ids`
to avoid running a `WP_Query` when there are no balancer articles.

Fixed a bug in articles block that wouldn't count correctly the amount of cells
displayed, thus showing the same articles that in previous rows.