- Added classes `TA_Balancer_Article` and `TA_Balancer_Author` to manage new database articles data in a way thats compatible with anything that uses `TA_Article_Data` or `TA_Author_Data` instances.
- In the Gutenberg Module, added the option `script_dependencies` in a
block's config to set the script dependencies if needed. This allows to enqueue
the script in places where the default dependencies that are used in the editor
aren't loaded.

- The row render got abstracted into `ta_render_articles_block_row`
to make it renderable outside the block by calling that function.

- Now balancer rows print a placeholder that will get replaced in the front
with a component that fetches balanced articles from the new database. This
fetch is yet to be balanced with the user balancer data.

- Added route `ta/v1/balancer-row` to retrieve a row's html based on the
params.

- Added route `ta/v1/balancer-db/articles` that fetches articles from the
new database.