Tags cloud module refactored. `ta/balancer-front` integrated with new tags cloud functions.
Tags cloud module script refactored and moved into `inc\gutenberg\src\blocks\ta-balancer-front\tagsCloud.js`.
It stores the data in a localStorage item named `taBalancerTagsCloud`.
This item has:
    - `ids`: The ids of the tags selected. If this field is populated, then
    the user complete the personalization.
    - `denied`:  Indicates if the user closed the tags cloud.

Now, selected tags in de module will get highlighted. The user is able to
remove a selected tag by clicking it again.

If the user clicks on the `listo` button, ids will be stored in the localStorage.

If the user closes the module, the module gets hidden and won't be shown again.

This data is sent to the balancer rows to request articles related to the stored
tags, if the user is anonymous.